Rune drop efficiency toa logger (#172)
* Guild Shop Runes and TOA Logger fix

When looking at and buying runes from the Guild Magic shop, it gets the rune drop efficiency like the regular magic shop.  The TOA logger has been fixed to correctly calculate the seconds.  Required and updated to the latest version of Electron to run correctly on Mojave so that’s in here too.

* reverting the electron version change

* Also reverting the json.lock file

* Ok, I think this takes care of the lock file